backend/requirements.txt
# FastAPI 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sse-starlette>=2.0.0

# LangChain 1.x
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langchain-deepseek>=0.1.0
langgraph>=0.2.0

# LlamaIndex
llama-index>=0.11.0
llama-index-core>=0.11.0
llama-index-embeddings-openai>=0.2.0
llama-index-retrievers-bm25>=0.3.0

# OpenAI Embedding
openai>=1.12.0

# 工具依赖
beautifulsoup4>=4.12.0
html2text>=2024.2.26
httpx>=0.26.0
aiohttp>=3.9.0

# Token 计数
tiktoken>=0.5.0

# YAML 解析
pyyaml>=6.0.1

# 环境变量
python-dotenv>=1.0.0
